Extract code to verify ISO document signature
