+++ vck/src/commonMain/kotlin/at/asitplus/wallet/lib/AbstractMdocVerifier.kt
@@ -1,29 +1,24 @@
 package at.asitplus.wallet.lib
 
 import at.asitplus.dcapi.SessionTranscriptContentHashable
-import at.asitplus.iso.DeviceAuthentication
 import at.asitplus.iso.Document
 import at.asitplus.iso.MobileSecurityObject
 import at.asitplus.iso.SessionTranscript
-import at.asitplus.iso.wrapInCborTag
-import at.asitplus.signum.indispensable.cosef.io.coseCompliantSerializer
 import at.asitplus.wallet.lib.agent.KeyMaterial
 import at.asitplus.wallet.lib.cbor.VerifyCoseSignatureWithKeyFun
-import io.matthewnelson.encoding.base16.Base16
-import io.matthewnelson.encoding.core.Encoder.Companion.encodeToString
-import kotlinx.serialization.encodeToByteArray
 
+@Deprecated("Logic moved to MdocDeviceSignatureVerifier")
 abstract class AbstractMdocVerifier {
     /** Creates challenges in authentication requests. */
     protected abstract val nonceService: NonceService
+
     /** Used for encrypted responses. */
     protected abstract val decryptionKeyMaterial: KeyMaterial
+
     /** Used to verify session transcripts from mDoc responses. */
     protected abstract val verifyCoseSignature: VerifyCoseSignatureWithKeyFun<ByteArray>
 
-    /**
-     * Performs calculation of the [at.asitplus.iso.SessionTranscript] for DC API
-     */
+    // Get's removed in 8.0.0
     protected abstract fun createDcApiSessionTranscript(
         toBeHashed: SessionTranscriptContentHashable,
     ): SessionTranscript
@@ -35,31 +30,8 @@ abstract class AbstractMdocVerifier {
     @Throws(IllegalArgumentException::class, IllegalStateException::class)
     protected fun verifyDocument(
         sessionTranscript: SessionTranscript
-    ): suspend (MobileSecurityObject, Document) -> Boolean = { mso, document ->
-        val deviceSignature = document.deviceSigned.deviceAuth.deviceSignature
-            ?: throw IllegalArgumentException("deviceSignature is null")
-
-        val expectedDetachedPayload = DeviceAuthentication(
-            type = DeviceAuthentication.TYPE,
-            sessionTranscript = sessionTranscript,
-            docType = document.docType,
-            namespaces = document.deviceSigned.namespaces
-        ).wrapAsExpectedPayload()
-
-        verifyCoseSignature(
-            coseSigned = deviceSignature,
-            signer = mso.deviceKeyInfo.deviceKey,
-            externalAad = byteArrayOf(),
-            detachedPayload = expectedDetachedPayload
-        ).onFailure {
-            throw IllegalArgumentException("deviceSignature not matching ${expectedDetachedPayload.encodeToString(Base16())}", it)
-        }
-        true
-    }
-
-    private fun DeviceAuthentication.wrapAsExpectedPayload(): ByteArray = coseCompliantSerializer
-        .encodeToByteArray(coseCompliantSerializer.encodeToByteArray(this))
-        .wrapInCborTag(24)
+    ): suspend (MobileSecurityObject, Document) -> Boolean =
+        MdocDeviceSignatureVerifier(verifyCoseSignature = verifyCoseSignature).verifyDocument(sessionTranscript)
 
 
 }
